pyproject.toml
[tool.poetry]
name = "ptback"
version = "0.1.0"
description = "Django as backend of portfolio tracker project"
authors = ["Julian Blacher <julian.blacher@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.2"
django-guardian = "^2.4.0"
django-filter = "^21.1"
django-markdown = "^0.8.4"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
